Migrate legacy mirror status page to Symfony controller and Twig template
